xxw/chaos/requirements.txt
# dev
black==18.6b4
isort==4.3.4
flake8==3.5.0
PyInstaller==3.3.1
pytest==3.6.3
pytest-cov==2.5.1
pytest-flake8==1.0.1

# web
Flask==1.0.2
Flask-RESTful==0.3.6
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
marshmallow-sqlalchemy==0.14.0
flask-marshmallow==0.9.0
PyMySQL==0.9.2
SQLAlchemy==1.2.10
kazoo==2.5.0
raven==6.9.0
requests==2.19.1
schema==0.6.8
redis==2.10.6
blinker==1.4

# server
gunicorn==19.8.1
meinheld==0.6.1

# third-part
qcloudsms-py==0.1.3
ipip-datx==0.4.0
